requirements.txt
langchain-community
openai
langchain_openai
tavily-python
streamlit
